Removing Netty from the public API
