+++ zio-http/src/main/scala/zio/http/Proxy.scala
@@ -26,30 +26,6 @@ final case class Proxy(
   def withUrl(url: URL): Proxy                         = self.copy(url = url)
   def withCredentials(credentials: Credentials): Proxy = self.copy(credentials = Some(credentials))
   def withHeaders(headers: Headers): Proxy             = self.copy(headers = headers)
-
-  /**
-   * Converts a Proxy to [io.netty.handler.proxy.HttpProxyHandler]
-   */
-  private[zio] def encode: Option[HttpProxyHandler] = credentials.fold(unauthorizedProxy)(authorizedProxy)
-
-  private def authorizedProxy(credentials: Credentials): Option[HttpProxyHandler] = for {
-    proxyAddress <- buildProxyAddress
-    uname          = credentials.uname
-    upassword      = credentials.upassword
-    encodedHeaders = headers.encode
-  } yield new HttpProxyHandler(proxyAddress, uname, upassword, encodedHeaders)
-
-  private def unauthorizedProxy: Option[HttpProxyHandler] = for {
-    proxyAddress <- buildProxyAddress
-    encodedHeaders = headers.encode
-  } yield {
-    new HttpProxyHandler(proxyAddress, encodedHeaders)
-  }
-
-  private def buildProxyAddress: Option[InetSocketAddress] = for {
-    proxyHost <- url.host
-    proxyPort <- url.port
-  } yield new InetSocketAddress(proxyHost, proxyPort)
 }
 
 object Proxy {
