+++ zio-http/src/main/scala/zio/http/netty/client/NettyClientDriver.scala
@@ -1,17 +1,20 @@
 package zio.http.netty.client
 
 import scala.collection.mutable
+
 import zio._
 import zio.stacktracer.TracingImplicits.disableAutoTrace
+
 import zio.http.ClientDriver.ChannelInterface
 import zio.http._
-import zio.http.netty.{ChannelFactories, EventLoopGroups, Names, NettyFutureExecutor, NettyRuntime, WebSocketAppHandler}
+import zio.http.netty._
+import zio.http.netty.model.Conversions
 import zio.http.socket.SocketApp
+
 import io.netty.channel.{Channel, ChannelFactory, ChannelHandler, EventLoopGroup}
 import io.netty.handler.codec.http.HttpObjectAggregator
 import io.netty.handler.codec.http.websocketx.WebSocketClientProtocolHandler
 import io.netty.handler.flow.FlowControlHandler
-import zio.http.netty.model.Conversions
 
 final case class NettyClientDriver private (
   channelFactory: ChannelFactory[Channel],
