+++ zio-http/src/test/scala/zio/http/netty/client/ClientRequestEncoderSpec.scala
@@ -1,12 +1,14 @@
 package zio.http.netty.client
 
-import io.netty.handler.codec.http.HttpHeaderNames
+import zio.test.Assertion._
+import zio.test._
+
 import zio.http.internal.HttpGen
 import zio.http.netty._
 import zio.http.netty.model.Conversions
 import zio.http.{Body, Request}
-import zio.test.Assertion._
-import zio.test._
+
+import io.netty.handler.codec.http.HttpHeaderNames
 
 object ClientRequestEncoderSpec extends ZIOSpecDefault with ClientRequestEncoder {
